Make view-models nested types of their view
