+++ ReadingList/BookList.swift
@@ -2,7 +2,7 @@ import SwiftUI
 
 struct BookList: View {
 
-    @ObservedObject var viewModel: BookListViewModel
+    @ObservedObject var viewModel: ViewModel
 
     var body: some View {
         List(viewModel.books) { item in
