--- ReadingList/ToReadListViewModel.swift
@@ -1,21 +1,0 @@
-import Combine
-
-class ToReadListViewModel: ObservableObject {
-
-    @Published var books: [BookToRead]
-
-    var readingListController: ReadingListController
-
-    private var cancellables = Set<AnyCancellable>()
-
-    init(readingListController: ReadingListController) {
-        self.readingListController = readingListController
-
-        books = readingListController.readingList
-
-        readingListController.$readingList.sink { [weak self] in
-            self?.books = $0
-        }
-        .store(in: &cancellables)
-    }
-}
